Results should be assigned first apparently...
When 'metadata' was assigned first, it was it that contained  the
response from the database and `results` was empty.
From sequelize docs:
''By default the function will return two arguments - a results array,
and an object containing metadata (affected rows etc.). Note that since
this is a raw query, the metadata (property names etc.) is dialect
specific. Some dialects return the metadata "within" the results object
(as properties on an array). However, two arguments will always be
returned, but for MSSQL and MySQL it will be two references to the same
object.''