:see_no_evil: Except that `credentials` can sometimes be singular...
Replace a bunch of instances of `credentials` with `connection` since
we were previously referring to `credentials` as a single object with a
few parameters. This is now just a `connection`.
Also, we were using `Connections.js` to refer to the different
Datastore backends. Rename this to a Datastore.js module.