include and display an error message
this is mostly for debugging and helping our users if something goes
wrong.
I think that the only mistake that a user could make that would cause
an error would be if they were using the desktop connector for
on-premise and they didn’t set their API domain correctly. in that
case, we the error message should be a little bit helpful - it’ll
indicate to them that the connector was reaching out to api.plot.ly
instead of `api-plotly.acme.com`